+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Use the SAM "Digest Update Multiple" command whenever possible.
+### Upgraded
+- "Calypsonet Terminal Calypso API" to version `1.3.0`.
 
 ## [2.2.1] - 2022-07-25
 ### Fixed